fix: switch assert to pure
Making assert elemental gives confusing output because only the
first element of array diagnost_data shows in the ouptut of a
failed assertion.  For printing array data as diagnostic output,
it is better for a user to package that data in an object and
pass a characterizable_t object as the actual diagnostic_data
argument.