CI: Check assertion output more rigorously
Extend steps that run assertion failures to check not only the expected
exit code, but also that the output appears to include a legible assertion
message. This is intended to catch cases where the ERROR STOP character
output is garbled, a defect we've now separately seen from two compilers.

New CHECK_ASSERT variable factors this logic and applies it where appropriate.

Force bash as the default shell, because otherwise containers default
to /bin/sh